Agent/RAG_智能客服与知识问答/app/config.py
import os
from dotenv import load_dotenv
load_dotenv()
QDRANT_URL = os.getenv("QDRANT_URL", "http://115.190.125.94:6333")
QDRANT_API_KEY = os.getenv("QDRANT_API_KEY","root")
QDRANT_COLLECTION = os.getenv("QDRANT_COLLECTION", "helpdesk-knowledge")

DASHSCOPE_API_KEY = os.getenv("DASHSCOPE_API_KEY")
DASHSCOPE_BASE_URL = os.getenv(
    "DASHSCOPE_BASE_URL",
    "https://dashscope.aliyuncs.com/compatible-mode/v1",
)

CHAT_MODEL = os.getenv("CHAT_MODEL", "qwen-plus")
EMBEDDING_MODEL = os.getenv("EMBEDDING_MODEL", "text-embedding-v4")
