Agent/RAG_智能客服与知识问答/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-text-splitters
langchain-community
langchain-qdrant
langchain-openai
qdrant-client
python-dotenv
tiktoken
openai
